chore(ecard): split demo
