+++ src/packages/__VUE/ecard/demo/index.vue
@@ -5,7 +5,7 @@
   </Demo>
 </template>
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 
 const t = useTranslate({
