+++ src/packages/__VUE/ecard/doc.en-US.md
@@ -16,51 +16,7 @@ app.use(Ecard);
 
 ### Basic
 
-:::demo
-
-```vue
-<template>
-  <nut-cell>
-    <nut-ecard
-      v-model="money"
-      @input-change="inputChange"
-      @change="change"
-      @change-step="changeStep"
-      :data-list="dataList"
-    ></nut-ecard>
-  </nut-cell>
-</template>
-<script setup>
-import { ref, reactive } from 'vue';
-const dataList = reactive([
-  {
-    price: 10
-  },
-  {
-    price: 20
-  },
-  {
-    price: 30
-  },
-  {
-    price: 40
-  }
-]);
-const money = ref(10);
-const inputChange = (val) => {
-  money.value = val;
-};
-const change = (item) => {
-  money.value = item.price;
-};
-const changeStep = (num, price) => {
-  const val = price * num;
-  money.value = val;
-};
-</script>
-```
-
-:::
+> demo: ecard basic
 
 ## API
 
@@ -68,7 +24,7 @@ const changeStep = (num, price) => {
 
 | Attribute | Description | Type | Default |
 | --- | --- | --- | --- |
-| modelValue | Price | number | `0` |
+| model-value | Price | number | `0` |
 | choose-text | Main Title | string | `Select` |
 | other-value-text | Other Text | string | `Other Value` |
 | data-list | Ecard List | Array | `DataList[]` |
